==> notebook/1_limpieza.ipynb <==
import pandas as pd
import numpy as np
import sys
import re

import seaborn as sns
import matplotlib.pyplot as plt

sys.path.append('../')
import src.limpieza as lmp

import requests

pd.options.display.max_columns = None

url = 'https://www.ine.es/jaxiT3/files/t/es/csv_bdsc/2078.csv'
res1 = requests.get(url)


res1.status_code

from pathlib import Path

data_folder = Path(r"C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data")
file_to_open = data_folder / "2078.csv"
with open(file_to_open, 'wb') as f:
    f.write(res1.content)

df_tourist = pd.read_csv(r"C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\2078.csv", sep=';', encoding='latin-1', index_col = 0)

df_tourist.shape

df_tourist.head(5)

df_tourist.duplicated().sum()

It seems that we have quite a lot of duplicates but, is it true?

df_tourist[df_tourist.duplicated() == True]

df_tourist[df_tourist['Total'] == '0']

Although it seemed that there were duplicates, in the previous table we can see that this is not the case and if we reset the index they disappear

df_tourist.reset_index(inplace=True)

df_tourist.duplicated().sum()

df_tourist.dtypes

We'll have to clean the column names

df_tourist.columns

print(df_tourist['Viajeros y pernoctaciones'].unique())
print(df_tourist['Residencia'].unique())

df_tourist.isnull().sum()

df_tourist.shape

We rename the columns

df_tourist = df_tourist.rename(columns={'ï»¿Puntos turÃ­sticos': 'Puntos_turisticos'})
columnas_limpias = {col : col.lower().replace(" ", "_") for col in df_tourist.columns}
df_tourist.rename(columns = columnas_limpias, inplace = True)

df_tourist.columns

Now let's clean up the 'total' column. We will first remove the dots and then convert the values ​​to integers.

It is important to note that in this csv if there is dot '.' alone, it may mean that the data is protected by the statistical secrecy law (Article 12 of Law 7/2006) or that the data was not collected. Both cases will be treated as if they were nulls.

This cleaning operation has therefore a double function:
- be able to convert strings to int
- convert the '.' in NaN to handle them later

df_tourist['total'] = df_tourist['total'].str.replace('.', '')

df_tourist['total'] = pd.to_numeric(df_tourist['total'], errors='coerce').astype('Int64')

We can see that the nulls went up from 5358 to 13056.

df_tourist.isnull().sum()

We know by the documentation of the csv that between april and june 2020 there is no data due to the covid-19 crisis. Let's take a look at that period.

df_tourist[(df_tourist['periodo'] == '2020M04') | (df_tourist['periodo'] == '2020M05') | (df_tourist['periodo'] == '2020M06')]

As during the covid-19 crisis all the turistic activity was suspended, we will substitute the nulls in those months by 0, as the problem wasn't in the data collection, but due to the fact that there were no tourist at all.

covid = ['2020M04', '2020M05', '2020M06']

df_tourist.loc[(df_tourist['periodo'].isin(covid)), 'total'] = 0

And we check that the nulls went down from 13056 to 12208

df_tourist.isnull().sum()

The nulls that we still have are due to two facts: there wasn't any data collection for place at those dates or the data is protected by statistic secretc law. Anyway, we cannot erase them directly; 
in the event that they are missing in specific months but we have data for the rest of the year, it may be convenient to replace them by the mean or the median.

null_rows = pd.isnull(df_tourist['total'])

df_tourist[null_rows].to_csv('nulls.csv')

df_tourist['periodo'] = df_tourist['periodo'].str.replace('M', '-')

df_tourist.to_pickle(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\tourist.pkl')
#df_tourist = pd.read_pickle(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\tourist.pkl')


df_tourist['periodo'] = pd.to_datetime(df_tourist['periodo'], format='%Y%m%')
df_tourist['periodo'] = df_tourist['periodo'].dt.date.apply(lambda x: x.strftime('%Y-%m'))

df_tourist.dtypes

df_tourist.set_index('periodo', inplace=True)

We rename the values

df_tourist['residencia'].replace({'Residentes en EspaÃ±a': "residentes_espana", 'Residentes en el Extranjero': 'residentes_en_el_extranjero'}, inplace=True)
df_tourist['viajeros_y_pernoctaciones'] = df_tourist['viajeros_y_pernoctaciones'].str.lower()

We apply the function that cleans and organize the data by cities

lista = ['Barcelona']

lmp.clean_tourist(df_tourist, lista)

df_tourist = pd.read_pickle(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\Barcelona_cleaned_tourist.pkl')

df_tourist.head(5)
