==> notebook/2_scraping.ipynb <==
import random
sys.path.append('../')
import src.scraping as scra

from time import sleep

from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.chrome.options import Options
from selenium.webdriver import ActionChains

import warnings
warnings.filterwarnings('ignore')

from selenium.webdriver.support import expected_conditions
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait

opciones= Options()
opciones.add_experimental_option('excludeSwitches', ['enable-automation'])
opciones.add_experimental_option('useAutomationExtension', False)
opciones.add_argument('--start-maximized')
opciones.add_argument('user.data-dir=selenium')
opciones.add_argument('--incognito')

url = 'https://www.civitatis.com/es/barcelona/free-tour-barcelona/opiniones/'

driver = webdriver.Chrome(ChromeDriverManager().install())

driver.get("https://www.civitatis.com/es/barcelona/free-tour-barcelona/opiniones/")
driver.implicitly_wait(10)

# Accept cookies
driver.find_element_by_css_selector("#didomi-notice-agree-button").click()

Now we will scrape all the reviews for the civitatis' free tour in Barcelona

contador = 0
scra.civitatiScrape()
