==> notebook/3_limpieza_scraping.ipynb <==
import re
import pandas as pd
sys.path.append('../')
import src.limpieza as lmp
import numpy as np

pd.options.display.max_columns = None

revius = open("output1.html", "r", encoding='utf-8').readlines()

civitatis_revius_limpio = lmp.revius_separados(revius)

longitud = [len(i) for i in civitatis_revius_limpio]
long_ = set(longitud)
print(long_)

for i in range(0,14):
    for a in civitatis_revius_limpio:
        if len(a)==i:
            print(a)
            break

civitatis_limpio_oi = lmp.pulir(civitatis_revius_limpio)
civitatis_reluciente = lmp.quitar_respuestas(civitatis_limpio_oi)


longitud = [len(i) for i in civitatis_reluciente]
long_ = set(longitud)
print(long_)

len(civitatis_reluciente)

dict_revius = lmp.dict_civitatis_completo(civitatis_reluciente)

len(dict_revius['fecha']) == len(dict_revius['nombre']) == len(dict_revius['procedencia']) == len(dict_revius['opinion']) == len(dict_revius['viajo_con'])

df_civitatis = pd.DataFrame(dict_revius)

df_civitatis.head(5)

df_civitatis.duplicated().sum()

We can se that we have quite a lot of duplicates, actually around 29% of the rows seem to be replicated but, is this completely truth?

df_civitatis[df_civitatis.duplicated() == True]

df_civitatis.iloc[13,3]

df_civitatis[df_civitatis['opinion'] == 'Casi 3 horas de tour pero Víctor, el guía que nos ha acompañado ha sido un encanto. Hemos andado muchísimo pero ha merecido la pena porque se nota que le gusta mucho su trabajo y se ha hecho muy ameno y emocionante. Sin duda un tour 100% recomendado con este guía. Gracias por hacernos pasar una mañana increíble.']


We can be absolutely sure that the we have doubled values only if we check the column 'opinion', as it's the only one unrepeatable. On the other hand, we can easely have several people from Spain doing a tour the same day who wouldn't leave their names, so all of them will appear as 'Anonimo' and 'España' with the other columns in blank, and pandas will take them as duplicates while they actually were different clients.
So let's analize this in two different data frames. First we'll analize the revius with the 'opinion' column filled.

dict_revius_5 = lmp.dict_civitatis_5(civitatis_reluciente)

df_civitatis_5 = pd.DataFrame(dict_revius_5)

df_civitatis_5.shape

df_civitatis_5.head(5)

df_civitatis_5.duplicated().sum()

df_civitatis_5[df_civitatis_5.duplicated() == True]

df_civitatis_5.iloc[0, 3]

df_civitatis_5[df_civitatis_5['opinion'] == ' Victor es encantador, te explica todo al detalle y te resuelve todas las preguntas, volvería a repetir, me a encantado.']

And here we can see that that some reviews are repeated up to three times... 
Anyway, from the total number of rows, 8953, this means that 23,6% of them are duplicated, a value significantly lower than the 29% we got for whole data frame.
Now let's drop the duplicates in this one and then we will work with the reviews that don't have the 'opinion' column filled.

df_civitatis_5 = df_civitatis_5.drop_duplicates()
df_civitatis_5.duplicated().sum()

df_civitatis_5[df_civitatis_5['opinion'] == ' Victor es encantador, te explica todo al detalle y te resuelve todas las preguntas, volvería a repetir, me a encantado.']

All the reviews whithout 'opinion':

dict_civitatis_2_4 = lmp.dict_civitatis_2_4(civitatis_reluciente)

df_civitatis_2_4 = pd.DataFrame(dict_civitatis_2_4)

df_civitatis_2_4.shape

df_civitatis_2_4.head(5)

df_civitatis_2_4.duplicated().sum()

If we take the reviews without 'opinion', we can see that the number of duplicates is signifincantly higher, around a 35%. This is due to the fact commented above that there can easely be two different clients with the same name (or name and origin) the same day.
Taking this into account, we won't drop all the duplicates, as this will alter significantly the data and we will be earasing real tourist from our data frame. But as we cannot state for sure which rows are duplicates and which ones are not, we will ramdomly erase 23.6% of them, as this seems to be the real percentage of duplicates that the algorithm generates; this means that we are going to drop 1717 rows in stead of the 2576 that pandas duplicated() method proposes.

df_civitatis_2_4 = lmp.drop_tourist(df_civitatis_2_4)

And we check that 23.6% of the rows are gone

df_civitatis_2_4.shape

Now that we have both datesets cleaned of duplicates, we can concatenate them

pd_civitatis_clean= pd.concat([df_civitatis_5, df_civitatis_2_4])

pd_civitatis_clean.shape

Above we verify that 23.6% of the rows are gone. From 16231 at the begining, now we have 12400.

Time to clean the data

pd_civitatis_clean.sample(5)

First, we will take away the spaces at the begining and end of the values

df_civitatis_clean = lmp.values_sin_espacios(df_civitatis)

Now we will deal with the dates ('fecha' column), so they will be in date-time format, as this will be the main criteria we will use to organize the tourist

df_civitatis_clean.iloc[0,0]

df_civitatis_clean = lmp.date_time(df_civitatis)

df_civitatis_clean.iloc[0,0]

df_civitatis_clean.sort_values(by=['fecha']).head(5)

df_civitatis_clean.to_pickle(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_clean.pkl')


df_turistas_fecha = df_civitatis_clean.set_index('fecha')

df_numero_turistas_fecha = df_turistas_fecha.groupby([(df_turistas_fecha.index.year),(df_turistas_fecha.index.month)]).count()

df_numero_turistas_fecha.tail(5)

df_numero_turistas_fecha = df_numero_turistas_fecha.drop(['procedencia', 'opinion', 'viajo_con'], axis=1).rename(columns={'nombre':'numero_turistas'})

df_numero_turistas_fecha.head(5)

df_numero_turistas_fecha.to_pickle(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\numero_turistas_fecha.pkl')
