==> notebook/4_dataframes.ipynb <==
In this notebook we will generate and save the data frames that we will use in visualization, using the clean ones done in the previous steps

import pandas as pd
sys.path.append('../')
import src.limpieza as lmp
from geopy.geocoders import Nominatim
import re
import geopandas as gpd
from keplergl import KeplerGl

df_civitatis_clean = pd.read_pickle(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_clean.pkl')
df_turistas_admon = pd.read_pickle(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\Barcelona_cleaned_tourist.pkl')


df_civitatis_location = pd.read_csv(r"C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_location.csv", sep=',', encoding='latin-1')

df_civitatis_location.tail(5)

# Civitatis

Total number of tourist grouped by month

lmp.total_civitatis_tourists_month(df_civitatis_clean)

df_civitatis_clean.shape

df_civitatis_clean.head(5)

Longitude and latitude of the place of origin of tourists

def extraer_coordenadas(x):
    try:
        geolocator = Nominatim(user_agent = "pepe")
        location = geolocator.geocode(x)

        return location.latitude, location.longitude
    except:
        return 'no data'

df_coordinates = pd.DataFrame(df_civitatis_clean['procedencia'].unique())

df_coordinates["georeferenced_column"] = df_coordinates[0].apply(extraer_coordenadas)

df_coordinates.to_pickle(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\lat_lon.pkl')


df_coordinates = pd.read_pickle(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\lat_lon.pkl')

df_coordinates.head(5)

def lat_lon(df, column):
    df['latitude'] = range(len(df))
    df['longitude'] = range(len(df))
    for i in range(len(df)):
        df['latitude'].iloc[i] = column.iloc[i][0]
        df['longitude'].iloc[i] = column.iloc[i][1]
    df['latitude'] = pd.to_numeric(df['latitude'],errors='coerce')
    df['longitude'] = pd.to_numeric(df['longitude'],errors='coerce')
    return df


df_civitatis_clean.head(5)

df_civitatis_clean.set_index(['procedencia'], inplace=True)
df_coordinates.set_index([0], inplace=True)

df_civitatis_location = df_civitatis_clean.join(df_coordinates, on='procedencia')

geolocator = Nominatim(user_agent="miguel")
location = geolocator.geocode("Barcelona, España")
df_civitatis_location['bar_lat'] = location.latitude
df_civitatis_location['bar_lon'] = location.longitude

df_civitatis_location['fecha'] = df_civitatis_location['fecha'].astype(str) + ' 00:00'

df_civitatis_location['fecha'] = pd.to_datetime(df_civitatis_location['fecha'], format='%Y-%m-%d %H:%M') 

df_civitatis_location['fecha']=df_civitatis_location['fecha'].dt.strftime('%Y-%m-%d %H:%M')

df_civitatis_location.dtypes

df_civitatis_location.to_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_location.csv')

df_civitatis_location_primary = pd.read_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_location_primary.csv')

df_civitatis_location_primary.drop(['Unnamed: 0'], axis=1, inplace=True)

df_num_loc =  df_civitatis_location_primary.groupby(['procedencia']).count()

df_num_loc.drop(['Unnamed: 0', 'fecha','georeferenced_column', 'latitude', 'longitude', 'periodo', 'opinion', 'viajo_con', 'bar_lat', 'bar_lon', 'primary_key'], axis=1, inplace=True)

df_num_loc.to_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civ_num_loc.csv')

df_num_loc.dtypes

df_num_loc['procedencia'].str.explode(',')

df_num_loc.reset_index(inplace=True)

df_oi = df_civitatis_location_primary.groupby(['periodo', 'georeferenced_column', 'latitude', 'longitude']).count()

df_oi.drop(['nombre','opinion', 'viajo_con', 'procedencia', 'bar_lat', 'bar_lon', 'primary_key'], axis=1, inplace=True)

df_oi['periodo'] = pd.to_datetime(df_oi['periodo'], format='%Y-%m-%d %H:%M') 

df_oi['periodo']=df_oi['periodo'].dt.strftime('%Y-%m-%d %H:%M')

df_oi.head(5)

df_oi.reset_index(inplace=True)

df_oi.to_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\oi.csv')

df_oi = pd.read_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\oi.csv')

map_1 = KeplerGl(height = 600, weight = 500, data = {"viajes": df_civitatis_location})

map_1.save_to_html(file_name='civitatis.html')

map_1

df_civitatis_location.head(5)

df_civitatis_location.drop(['Unnamed: 0'], axis=1, inplace=True)

# Spanish goverment

As civitatis started its activity for the tour that we are studying in June 2017, we'll select that period from the data frame we had generated using the data given by the goverment.


lmp.admon_cividates(df_turistas_admon)

df_turistas_admon_17 = pd.read_pickle(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\admon_cividates.pkl')

df_turistas_admon_17.head(5)
