==> notebook/6_sql.ipynb <==
import pandas as pd
pd.options.display.max_columns = None

import sqlalchemy as alch
from sqlalchemy.exc import SQLAlchemyError

from getpass import getpass

df_turistas_admon_17 = pd.read_pickle(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\admon_cividates.pkl')
df_civitatis_location = pd.read_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_location.csv')

First we're going to create the column that will be the primary key in the sql database. We will set the primary key based on the date by months

df_turistas_admon_17.head(5)

df_turistas_admon_17['primary_key'] = df_turistas_admon_17.index

df_turistas_admon_primary = df_turistas_admon_17
df_turistas_admon_primary.to_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\turistas_admon_primary.csv')

df_turistas_admon_primary = pd.read_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\turistas_admon_primary.csv')

df_civitatis_location.head(1)

df_civitatis_location['periodo'] = pd.to_datetime(df_civitatis_location['fecha'], format='%Y-%m')
df_civitatis_location['periodo'] = df_civitatis_location['periodo'].dt.strftime('%Y-%m')

df_primary_key = df_turistas_admon_primary[['periodo', 'primary_key']]

df_primary_key.to_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\primary_key.csv')

df_primary_key.sort_values(by=['periodo', 'primary_key'], ascending=False, inplace=True)

df_primary_key.head(5)

df_primary_key.set_index(['periodo'], inplace=True)
df_civitatis_location.set_index(['periodo'], inplace=True)

df_civitatis_location_primary = df_civitatis_location.merge(df_primary_key, on='periodo')

df_civitatis_location_primary.drop(['Unnamed: 0'], axis=1, inplace=True)

df_civitatis_location_primary.tail(5)

df_civitatis_location_primary.to_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_location_primary.csv')

df_civitatis_location_primary = pd.read_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_location_primary.csv')

class Cargar:
    
    def __init__(self, nombre_bbdd, contraseña):

        # nuestra clase va a recibir dos parámetros que son fijos a lo largo de toda la BBDD, el nombre de la BBDD y la contraseña con el servidor. 
        self.nombre_bbdd = nombre_bbdd
        self.contraseña = contraseña

    def conexion_servidor(self): 
        conexion = f"mysql+pymysql://root:{self.contraseña}@localhost"
        return alch.create_engine(conexion)


    def crear_bbdd(self):
        engine = self.conexion_servidor()
        try:
            engine.execute(f"CREATE DATABASE IF NOT EXISTS {self.nombre_bbdd};")
            
        except:
            print("La BBDD ya existe")
            
    def crear_schema(self):
        engine = self.conexion_servidor()
        try:
            engine.execute(f"CREATE SCHEMA IF NOT EXISTS {self.nombre_bbdd} DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_bin;")
        except:
            print("El schema ya existe")

    def conexion_base_datos(self):

        conexion2 = f"mysql+pymysql://root:{self.contraseña}@localhost/{self.nombre_bbdd}"
        return alch.create_engine(conexion2)

    def crear_insertar_tabla(self, query):
        engine = self.conexion_base_datos()

    
        try:
            engine.execute(query)
          
        except SQLAlchemyError as e:
            error = str(e.__dict__['orig'])
            return error

contraseña = getpass("Contraseña de MySQL: ")

civitatis_sql = Cargar("turismo_admon_civitatis", contraseña)
civitatis_sql.crear_schema()

df_turistas_admon_primary.head(1)

table_admon_esp = '''
CREATE TABLE IF NOT EXISTS `turismo_admon_civitatis`.`admon_esp` (
        `clave` int(11) NOT NULL,
        `periodo` varchar(255) NOT NULL,
        `viajeros_espana` int(11) NOT NULL,
        `pernoctaciones_espana` int(11) NOT NULL,
        `media_pernoc_espana` FLOAT NOT NULL,
    PRIMARY KEY (`clave`));
'''

table_admon_ext = '''
CREATE TABLE IF NOT EXISTS `turismo_admon_civitatis`.`admon_ext` (
        `clave` int(11) NOT NULL,
        `periodo` varchar(255) NOT NULL,
        `viajeros_extranjero` int(11) NOT NULL,
        `pernoctaciones_extranjero` int(11) NOT NULL,
        `media_pernoc_extranjero` FLOAT NOT NULL,
    PRIMARY KEY (`clave`));
'''

table_civitatis_esp = '''
CREATE TABLE IF NOT EXISTS `turismo_admon_civitatis`.`civitatis_esp` (
        `clave` int(11) NOT NULL,
        `periodo` varchar(255) NOT NULL,
        `procedencia` VARCHAR(255),
        `fecha` DATE NOT NULL,
        `nombre` VARCHAR(255),
        `opinion` VARCHAR(255),
        `viajo_con` VARCHAR(255),
        `georeferenced_column` VARCHAR(255),
        `latitude` FLOAT,
        `longitude` FLOAT,
        `bar_latitude` FLOAT,
        `bar_longitude` FLOAT,
    PRIMARY KEY (`clave`));
'''

table_civitatis_ext = '''
CREATE TABLE IF NOT EXISTS `turismo_admon_civitatis`.`civitatis_ext` (
        `clave` int(11) NOT NULL,
        `periodo` varchar(255) NOT NULL,
        `procedencia` VARCHAR(255),
        `fecha` DATE NOT NULL,
        `nombre` VARCHAR(255),
        `opinion` VARCHAR(255),
        `viajo_con` VARCHAR(255),
        `georeferenced_column` VARCHAR(255),
        `latitude` FLOAT,
        `longitude` FLOAT,
        `bar_latitude` FLOAT,
        `bar_longitude` FLOAT,
    PRIMARY KEY (`clave`));
'''

table_total_turistas = '''
CREATE TABLE IF NOT EXISTS `turismo_admon_civitatis`.`total_turistas` (
        `clave` int(11) NOT NULL,
        `periodo` varchar(255) NOT NULL,
        `total_viajeros` int(11) NOT NULL,
        `total_pernoctaciones` int(11) NOT NULL,
        `media_total_pernoctaciones` FLOAT NOT NULL,
         PRIMARY KEY (`clave`, `admon_esp_clave`, `admon_ext_clave`, `civitatis_ext_clave`, `civitatis_esp_clave`),
  INDEX `fk_total_turistas_admon_esp_idx` (`admon_esp_clave` ASC) VISIBLE,
  INDEX `fk_total_turistas_admon_ext1_idx` (`admon_ext_clave` ASC) VISIBLE,
  INDEX `fk_total_turistas_civitatis_ext1_idx` (`civitatis_ext_clave` ASC) VISIBLE,
  INDEX `fk_total_turistas_civitatis_esp1_idx` (`civitatis_esp_clave` ASC) VISIBLE,
  CONSTRAINT `fk_total_turistas_admon_esp`
    FOREIGN KEY (`admon_esp_clave`)
    REFERENCES `turismo_admon_civitatis`.`admon_esp` (`clave`),
  CONSTRAINT `fk_total_turistas_admon_ext1`
    FOREIGN KEY (`admon_ext_clave`)
    REFERENCES `turismo_admon_civitatis`.`admon_ext` (`clave`),
  CONSTRAINT `fk_total_turistas_civitatis_ext1`
    FOREIGN KEY (`civitatis_ext_clave`)
    REFERENCES `turismo_admon_civitatis`.`civitatis_ext` (`clave`),
  CONSTRAINT `fk_total_turistas_civitatis_esp1`
    FOREIGN KEY (`civitatis_esp_clave`)
    REFERENCES `turismo_admon_civitatis`.`civitatis_esp` (`clave`))
ENGINE = InnoDB
DEFAULT CHARACTER SET = utf8mb4
COLLATE = utf8mb4_bin);
'''

civitatis_sql.crear_insertar_tabla(table_admon_esp)
civitatis_sql.crear_insertar_tabla(table_admon_ext)
civitatis_sql.crear_insertar_tabla(table_civitatis_esp)
civitatis_sql.crear_insertar_tabla(table_civitatis_ext)
#civitatis_sql.crear_insertar_tabla(table_total_turistas)

df_civitatis_location_primary.sample(1)

df_civitatis_location_primary_null = df_civitatis_location_primary[df_civitatis_location_primary['procedencia'].isnull()]

df_civitatis_location_primary_null.to_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_location_primary_null.csv')

df_civitatis_location_primary_null.shape

df_civitatis_location_primary_esp = df_civitatis_location_primary[df_civitatis_location_primary['procedencia'].str.contains('España') == True]


df_civitatis_location_primary_esp.to_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_location_primary_esp.csv')

df_civitatis_location_primary_esp.shape

df_civitatis_location_primary_ext = (df_civitatis_location_primary[df_civitatis_location_primary['procedencia'].str.contains('España') == False])

df_civitatis_location_primary_ext.to_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_location_primary_ext.csv')

df_civitatis_location_primary_ext.shape

df_civitatis_location_primary_ext.isnull().sum()

df_civitatis_location_primary_esp = pd.read_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_location_primary_esp.csv')
df_civitatis_location_primary_ext = pd.read_csv(r'C:\Users\mituc\Ironhack\Curso\IronLabs\Proyecto-1\data\civitatis_location_primary_ext.csv')

# Insert values in admons_esp
for indice, fila in df_turistas_admon_primary.iterrows():
    
    query_admon_esp = f"""INSERT INTO admon_esp (clave, periodo, viajeros_espana, pernoctaciones_espana, media_pernoc_espana) 
                        VALUES ({fila['primary_key']}, '{fila['periodo']}', {fila['viajeros_espana']}, {fila['pernoctaciones_espana']}, {fila['media_pernoc_esp']})"""

    civitatis_sql.crear_insertar_tabla(query_admon_esp)

# Insert values in admons_ext
for indice, fila in df_turistas_admon_primary.iterrows():
    
    query_admon_ext = f"""INSERT INTO admon_ext (clave, periodo, viajeros_extranjero, pernoctaciones_extranjero, media_pernoc_extranjero) 
                        VALUES ({fila['primary_key']}, '{fila['periodo']}', {fila['viajeros_extranjero']}, {fila['pernoctaciones_extranjero']}, {fila['media_pernoc_ext']})"""

    civitatis_sql.crear_insertar_tabla(query_admon_ext)

# Insert values in civitatis_esp
for indice, fila in df_civitatis_location_primary_esp.iterrows():
    
    query_civitatis_esp = f"""INSERT INTO civitatis_esp (clave, periodo, procedencia, fecha, nombre, opinion, viajo_con, georeferenced_column, latitude, longitude, bar_latitude, bar_longitude)) 
                        VALUES ({fila['primary_key']}, '{fila['periodo']}', '{fila['procedencia']}', {fila['fecha']}, {fila['nombre']}, {fila['opinion']}, {fila['viajo_con']}, {fila['georeferenced_column']}, {fila['latitude']}, {fila['longitude']}, {fila['bar_lat']}, {fila['bar_lon']})"""

    try:
        civitatis_sql.crear_insertar_tabla(query_civitatis_esp)
    except:
        print(f'Error en la insercion de la fila {fila}')
        continue

# Insert values in civitatis_esp
for indice, fila in df_civitatis_location_primary_ext.iterrows():
    
    query_civitatis_ext = f"""INSERT INTO civitatis_ext (clave, periodo, procedencia, fecha, nombre, opinion, viajo_con, georeferenced_column, latitude, longitude, bar_latitude, bar_longitude)) 
                        VALUES ({fila['primary_key']}, '{fila['periodo']}', '{fila['procedencia']}', {fila['fecha']}, {fila['nombre']}, {fila['opinion']}, {fila['viajo_con']}, {fila['georeferenced_column']}, {fila['latitude']}, {fila['longitude']}, {fila['bar_lat']}, {fila['bar_lon']})"""

    try:
        civitatis_sql.crear_insertar_tabla(query_civitatis_ext)
    except:
        print(f'Error en la insercion de la fila {fila}')
        continue

# Insert values in total_turistas
for indice, fila in df_turistas_admon_primary.iterrows():
    
    query_admon_total = f"""INSERT INTO total_turistas (clave, periodo, viajeros_extranjero, total_pernoctaciones, media_total_pernoctaciones) 
                        VALUES ({fila['primary_key']}, '{fila['periodo']}', {fila['total_viajeros']}, {fila['total_pernoctaciones']}, {fila['media_total_pernoc']})"""

    civitatis_sql.crear_insertar_tabla(query_admon_total)
